refactor: move editorship tests to their own module
